# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ecom API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Provides an overview of the admin's dashboard data, including recent sales, total sales, and product counts for the last 30 days.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get admin dashboard data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/checkoutcod": {
            "post": {
                "description": "Process a cash-on-delivery (COD) order and create payment, order, and order items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Process a COD order",
                "parameters": [
                    {
                        "description": "Order object with Address_ID",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully placed the order",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error in processing the order",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/deletefromcart/:cart_id": {
            "delete": {
                "description": "Delete a product from the user's cart by specifying the cart item ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Delete a product from the user's cart.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart Item ID",
                        "name": "cart_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted from cart",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/success": {
            "get": {
                "description": "Displays a success page after payment completion.",
                "produces": [
                    "text/html"
                ],
                "summary": "Payment Success",
                "operationId": "Success",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment success page",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error while displaying success page",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/viewcart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the list of products in the user's shopping cart.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Retrieve the user's shopping cart items.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cart items and total price",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Order": {
            "type": "object",
            "properties": {
                "addressid": {
                    "type": "integer"
                },
                "orderid": {
                    "type": "integer"
                },
                "paymentid": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "totalprice": {
                    "type": "integer"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
...
    }
}